Run Athena Setup

Stepwise process of instantiating the Athena Vocabularies. Steps can be skipped using the steps argument though the order cannot be changed.

run_setup(
  conn,
  conn_fun,
  target_schema = "omop_vocabulary",
  steps = c("drop_tables", "copy", "indices", "constraints", "log"),
  path_to_csvs,
  release_version,
  umls_api_key,
  verbose = TRUE,
  render_sql = TRUE
)

Arguments

conn
A Postgres connection object from the DatabaseConnector package. The user must either be a superuser or a member of the pg_read_server_files role.

conn_fun
Character string of the expression used to connect to the target database for the setup. The user must either be a superuser or a member of the pg_read_server_files role.This option automatically closes the connection on exit.

target_schema
Schema where the OMOP Vocabulary Tables will be written to.

steps
Options include c("prepare_cpt4", "drop_tables","copy", "indices", "constraints", "log"). Each step corresponds to the function used and may be independently executed. prepare_cpt4 is not included in default because of its tendency to run slow and the java script used to reconstitute the CPT4 needs to run uninterrupted or else the resulting CONCEPT.csv file will be damaged. Therefore, it is advised to reconstitute CPT4 separately from the Command Line or using the aforementioned function in this package though the option to do it as part of run_setup is available.

path_to_csvs
Path to folder containing the unpacked vocabularies as csv files, downloaded as a zip from athena.ohdsi.org.

release_version
Release version as an exact string provided vocabulary download link delivery via email.

umls_api_key
UMLS API Key for CPT4 reconstitution.

verbose
If TRUE, prints back activity in the R console as it is occurring.

render_sql
If TRUE, the SQL statement is printed back in the R console for reference.

Value

Updated OMOP Vocabulary (Athena) CONCEPT_ANCESTOR, C ONCEPT_CLASS, CONCEPT_RELATIONSHIP, CONCEPT_SYNONYM, CONCEPT, DOMAIN, DRUG_STRENGTH, RELATIONSHIP, and VOCABULARY Tables in the given target schema.

See also